backend/requirements.txt
fastapi==0.115.5
psycopg2-binary==2.9.10
pydantic==2.10.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn==0.32.1